+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Python: Added APPEND command ([#1152](https://github.com/aws/glide-for-redis/pull/1152))
 * Python: Added GEOADD command ([#1259](https://github.com/aws/glide-for-redis/pull/1259))
 * Python: Added GEOHASH command ([#1281](https://github.com/aws/glide-for-redis/pull/1281))
+* Python: Added ZLEXCOUNT command ([#1305](https://github.com/aws/glide-for-redis/pull/1305))
 
 #### Fixes
 * Python: Fix typing error "‘type’ object is not subscriptable" ([#1203](https://github.com/aws/glide-for-redis/pull/1203))